# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Server pets",
        "title": "Swagger Example API",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
...
        "/Node/createNginxNode": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "创建NginxNode",
                "parameters": [
                    {
                        "description": "创建NginxNode",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NginxNode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Node/deleteNginxNode": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "删除NginxNode",
                "parameters": [
                    {
                        "description": "删除NginxNode",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NginxNode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Node/deleteNginxNodeByIds": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "批量删除NginxNode",
                "parameters": [
                    {
                        "description": "批量删除NginxNode",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.IdsReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"批量删除成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Node/findNginxNode": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "用id查询NginxNode",
                "parameters": [
                    {
                        "description": "用id查询NginxNode",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NginxNode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"查询成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Node/getNginxNodeList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "获取NginxNode列表",
                "parameters": [
                    {
                        "description": "获取NginxNode列表",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NginxNodeSearch"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/Node/updateNginxNode": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "NginxNode"
                ],
                "summary": "更新NginxNode",
                "parameters": [
                    {
                        "description": "更新NginxNode",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.NginxNode"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"更新成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/autoCode/createTemp": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "自动代码模板",
                "parameters": [
                    {
                        "description": "创建自动代码",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.AutoCodeStruct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"创建成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autoCode/getColumn": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前表所有字段",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autoCode/getDatabase": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前所有数据库",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/autoCode/getTables": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "AutoCode"
                ],
                "summary": "获取当前数据库所有表",
                "responses": {
                    "200": {
                        "description": "{\"success\":true,\"data\":{},\"msg\":\"获取成功\"}",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.AutoCodeStruct": {
            "type": "object",
            "properties": {
                "abbreviation": {
                    "type": "string"
                },
                "autoConfig": {
                    "$ref": "#/definitions/model.AutoConfigStruct"
                },
                "autoCreateApiToSql": {
                    "type": "boolean"
                },
                "autoMoveFile": {
                    "type": "boolean"
                },
                "description": {
                    "type": "string"
                },
                "fields": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Field"
                    }
                },
                "packageName": {
                    "type": "string"
                },
                "structName": {
                    "type": "string"
                },
                "tableName": {
                    "type": "string"
                }
            }
        },
        "model.AutoConfigStruct": {
            "type": "object",
            "properties": {
                "Domain": {
                    "$ref": "#/definitions/model.NginxDomain"
                },
                "Pools": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.PoolConfig"
                    }
                },
                "Rules": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.RuleConfig"
                    }
                }
            }
        },
        "model.DomainRule": {
            "type": "object",
            "properties": {
                "Domain": {
                    "$ref": "#/definitions/model.NginxDomain"
                },
                "DomainId": {
                    "type": "integer"
                },
                "HttpsRewrite": {
                    "type": "string"
                },
                "MatchType": {
                    "type": "string"
                },
                "Path": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.DomainRuleIns": {
            "type": "object",
            "properties": {
                "InsType": {
                    "type": "string"
                },
                "InsValue": {
                    "type": "string"
                },
                "Rule": {
                    "$ref": "#/definitions/model.DomainRule"
                },
                "RuleId": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.Field": {
            "type": "object",
            "properties": {
                "columnName": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "dataType": {
                    "type": "string"
                },
                "dataTypeLong": {
                    "type": "string"
                },
                "dictType": {
                    "type": "string"
                },
                "fieldDesc": {
                    "type": "string"
                },
                "fieldJson": {
                    "type": "string"
                },
                "fieldName": {
                    "type": "string"
                },
                "fieldSearchType": {
                    "type": "string"
                },
                "fieldType": {
                    "type": "string"
                }
            }
        },
        "model.NginxCluster": {
            "type": "object",
            "properties": {
                "ClusterName": {
                    "type": "string"
                },
                "Department": {
                    "type": "string"
                },
                "Vip": {
                    "type": "string"
                },
                "authority": {
                    "$ref": "#/definitions/model.SysAuthority"
                },
                "authorityId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.NginxDomain": {
            "type": "object",
            "properties": {
                "Domain": {
                    "type": "string"
                },
                "DomainName": {
                    "type": "string"
                },
                "HttpsCert": {
                    "type": "string"
                },
                "HttpsPort": {
                    "type": "string"
                },
                "HttpsRewrite": {
                    "type": "string"
                },
                "Nginx": {
                    "type": "integer"
                },
                "NginxCluster": {
                    "$ref": "#/definitions/model.NginxPool"
                },
                "Pool": {
                    "type": "integer"
                },
                "PoolCluster": {
                    "$ref": "#/definitions/model.NginxPool"
                },
                "Port": {
                    "type": "string"
                },
                "UseHttps": {
                    "type": "boolean"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.NginxNode": {
            "type": "object",
            "properties": {
                "Cluster": {
                    "$ref": "#/definitions/model.NginxCluster"
                },
                "ClusterId": {
                    "type": "integer"
                },
                "IP": {
                    "type": "string"
                },
                "NodeName": {
                    "type": "string"
                },
                "Passwd": {
                    "type": "string"
                },
                "Port": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.NginxPool": {
            "type": "object",
            "properties": {
                "CheckInterval": {
                    "type": "integer"
                },
                "CheckTimeout": {
                    "type": "integer"
                },
                "CheckType": {
                    "type": "string"
                },
                "Keepalive": {
                    "type": "integer"
                },
                "Policy": {
                    "type": "string"
                },
                "PoolName": {
                    "type": "string"
                },
                "authority": {
                    "$ref": "#/definitions/model.SysAuthority"
                },
                "authorityId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.PoolConfig": {
            "type": "object",
            "properties": {
                "Pool": {
                    "$ref": "#/definitions/model.NginxPool"
                },
                "PoolNodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PoolNode"
                    }
                }
            }
        },
        "model.PoolNode": {
            "type": "object",
            "properties": {
                "Cluster": {
                    "$ref": "#/definitions/model.NginxPool"
                },
                "ClusterId": {
                    "type": "integer"
                },
                "IP": {
                    "type": "string"
                },
                "MaxFailed": {
                    "type": "string"
                },
                "NodeName": {
                    "type": "string"
                },
                "Port": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "TimeOut": {
                    "type": "string"
                },
                "Weight": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "model.RuleConfig": {
            "type": "object",
            "properties": {
                "Ins": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.DomainRuleIns"
                    }
                },
                "Pool": {
                    "$ref": "#/definitions/model.NginxPool"
                },
                "Rule": {
                    "$ref": "#/definitions/model.DomainRule"
                }
            }
        },
...
        "model.SysAuthority": {
            "type": "object",
            "properties": {
                "authorityId": {
                    "type": "string"
                },
                "authorityName": {
                    "type": "string"
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "dataAuthorityId": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "deletedAt": {
                    "type": "string"
                },
                "menus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                },
                "parentId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.SysBaseMenu": {
            "type": "object",
            "properties": {
                "authoritys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysAuthority"
                    }
                },
                "children": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenu"
                    }
                },
                "component": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "defaultMenu": {
                    "type": "boolean"
                },
                "hidden": {
                    "type": "boolean"
                },
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "keepAlive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SysBaseMenuParameter"
                    }
                },
                "parentId": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "sort": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "model.SysBaseMenuParameter": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "key": {
                    "type": "string"
                },
                "sysBaseMenuID": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "request.IdsReq": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "request.NginxNodeSearch": {
            "type": "object",
            "properties": {
                "Cluster": {
                    "$ref": "#/definitions/model.NginxCluster"
                },
                "ClusterId": {
                    "type": "integer"
                },
                "IP": {
                    "type": "string"
                },
                "NodeName": {
                    "type": "string"
                },
                "Passwd": {
                    "type": "string"
                },
                "Port": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "UserName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "x-token",
            "in": "header"
        }
    }
}